requirements.txt
Flask
gunicorn
grpcio
grpcio-tools
protobuf
boto3
botocore
torch
torchvision
pillow
numpy
typing
scikit-image
huggingface_hub
